pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-giffer"
version = "0.1.0"
description = "Create a time-lapse GIF from portrait selfies with face alignment"
requires-python = ">=3.10,<3.13"
dependencies = [
    "opencv-python>=4.8.0",
    "dlib>=19.24",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.7.1",
    "imageio>=2.31.0",
    "gdown>=5.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools]
packages = ["src"]
py-modules = ["main"]

